feat(guides): consolidate test-grader and run-grader into unified pnpm grade command
